Rename submodule to accounting
